Move content of orderedbytesmap.go to btrie.go.
